// src/app/shared/api-client/model/ingredientResponse.ts
/**
 * Pizza.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { IngredientType } from './ingredientType';


export interface IngredientResponse { 
    price?: number;
    id?: number;
    name?: string | null;
    ingredientType?: IngredientType;
}