// src/app/shared/api-client/model/ingredientType.ts
/**
 * Pizza.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type IngredientType = 'sauce' | 'cheese' | 'topping';

export const IngredientType = {
    Sauce: 'sauce' as IngredientType,
    Cheese: 'cheese' as IngredientType,
    Topping: 'topping' as IngredientType
};