// src/app/shared/api-client/model/size.ts
/**
 * Pizza.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type Size = 'small' | 'medium' | 'large';

export const Size = {
    Small: 'small' as Size,
    Medium: 'medium' as Size,
    Large: 'large' as Size
};